backend/requirements.txt
# --- Core web backend ---
fastapi==0.115.6
uvicorn==0.32.1
starlette==0.41.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
anyio==4.10.0
httpx==0.28.1
orjson==3.11.3

# --- LLMs & agents ---
openai==1.106.1           # pour Mistral via OpenRouter ou autres
anthropic==0.68.0         # si tu veux tester Claude / autres
langchain==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.33
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.23

# --- RAG & vecteurs ---
qdrant-client==1.12.1     # (ajouté, pas dans ta liste)
tiktoken==0.11.0

# --- Google GenAI (embeddings, etc.) ---
google-genai==1.34.0
google-ai-generativelanguage==0.9.0
google-api-core==2.25.1
google-auth==2.40.3

# --- Supabase ---
supabase==2.20.0
gotrue==2.12.3
postgrest==1.1.1
storage3==0.12.0
realtime==2.6.0

# --- Scraping / CrawlAI integration helpers ---
requests==2.32.5
beautifulsoup4==4.13.5

# --- Clustering / ML (plus tard) ---
numpy==2.3.2
pandas==2.3.3
scikit-learn==1.7.2
hdbscan==0.8.40
umap-learn==0.5.9.post2

# --- Redis (si tu fais la queue plus tard) ---
redis==6.4.0
rq==1.16.2   # ajout pour Redis Queue simple

# --- Dev / tests ---
pytest==7.4.3
pytest-asyncio==0.21.1
